c: fix sub-optimal skip
When performing skip scanning we cut our jump length in half if we hit
something and continue to do so until we home in on it. What we forgot
to do was reset the skip when we get a negative hit because that
indicates we've reset our scan state.

It only cuts the number of byte comparisons in half, no big deal.