c: use perfect sub-length ordering
Based on the counts we produce for the main sample file, order the
byte-by-byte checks in the matching function to optimize for most likely
to early exit. This optimization works by checking the fewest characters
possible before switching us to the long range skip mode.

The order won't be perfect for every input but should be reasonable
overall.

This is an obvious case of optimizing for a benchmark, but that is kind
of the goal of the exercise a this point since we've gone way beyond the
"simplest solution that works".