rust: reduce read buffer size
Reduce the size of the read buffer from 64MB to 256K which is much
friendlier to the CPU cache. This is the same size as used by the
optimized C version and it turns out to yield a small performance boost
here as well.

In fact, it provides enough of a speed up that the naive implementation
in Rust is now consistently faster than ripgrep, which is a little
surprising.