docs: refresh baseline times
These changes reflect a combination of newer hardware and 3 years of
improvements in the various compilers and runtimes.

Most notably, the Go implementation now runs in similar time to ripgrep.
It now takes 5 times longer than the C implementation instead of 20
times longer.

Another huge improvement comes from Node.js, which doubled in speed.
It's still nowhere near the compiled implementations, but it's a lot
closer than it used to be.

There's a small improvement in the rust version, but it looks like at
least half of the improvement is from running on faster hardware.

As with the original results, these numbers aren't meant to be precise
or reproducible and are really only meant for broadly comparing the
performance of the different implementations.