rust: add basic skipping
The switch from enumberate() to a regular loop is actually slower for
the single-stepping case, but the performance gains from being able to
skip large chunks of data with minimal comparisons more than makes up
for it.

This is the most basic optimization for this type of algorithm and this
is the simplified version of it. If we're not in a match, check 20
characters ahead and if that also isn't a match, skip 20. If it is a
match, then just continue on with checking each byte.