node: add basic skipping
The mighty skip comes to node, putting it nearly in the range of the
compiled languages.

Even though it isn't as fast as the compiled languages, this is still
objectively fast. Unless the file is already cached in memory, you're
well into the realm of being i/o bound instead of node being the
bottleneck.